+++ pkg/twitter/post_tweet.go
@@ -1,0 +1,40 @@
+package twitter
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/ganyariya/misskey-twitter-post/pkg/misskey"
+	"github.com/michimani/gotwi"
+	"github.com/michimani/gotwi/tweet/managetweet"
+	"github.com/michimani/gotwi/tweet/managetweet/types"
+)
+
+func PostToTwitter(misskeyRequest *misskey.MisskeyRequest) error {
+
+	c, err := gotwi.NewClient(&gotwi.NewClientInput{
+		AuthenticationMethod: gotwi.AuthenMethodOAuth1UserContext,
+		OAuthToken:           os.Getenv("USER_TWITTER_OAUTH_ACCESS_TOKEN"),
+		OAuthTokenSecret:     os.Getenv("USER_TWITTER_OAUTH_ACCESS_TOKEN_SECRET"),
+	})
+
+	if err != nil {
+		log.Println("Twitter Client Init Error: ", err.Error())
+		return err
+	}
+
+	misskeyAccountUrl := os.Getenv("MISSKEY_ACCOUNT")
+	p := &types.CreateInput{
+		Text: gotwi.String(fmt.Sprintf("%s %s", misskeyRequest.Body.Note.Text, misskeyAccountUrl)),
+	}
+
+	res, err := managetweet.Create(context.Background(), c, p)
+	if err != nil {
+		log.Println("Twitter Post Error: ", err.Error())
+		return err
+	}
+	log.Printf("[%s] %s\n", gotwi.StringValue(res.Data.ID), gotwi.StringValue(res.Data.Text))
+	return nil
+}
